Use rollup replace plugin to remove some node logic from the browser bundle on compilation
We had some downstream pain consuming this library in a bundled web app, due to this code:

```
if (typeof window === 'undefined' &&
        typeof require === 'function' &&
        require('fs') &&
        param instanceof require('fs').ReadStream) {
    return true;
}
```

Although the `require('fs')` doesn't actually get run, webpack was convinced that 'fs' was a dependency when trying to build, due to the require('fs') import.

We can work around this on our end, but it's nice if the web bundle just doesn't contain that code at all, so this uses a rollup replacement to remove it.