Update Configuration imports to ES6 (#784)
Reported in https://developer.genesys.cloud/forum/t/cant-use-purecloud-platform-client-v2-with-node-js-typescript/12151/10

> ReferenceError: You are trying to import a file after the Jest environment has been torn down

This warning reveals itself when live reloading is enabled and the context of the runtime is Node.js (not a browser). Recommended resolutions did not appear to have an effect on resolving the warning. The warning leads to having an impact on the success of Jest executions in CI/CD tooling.

In this commit:
- Update configuration.js to use ES6 imports. Removes runtime ESM require() from functions. Including the TS config for allowSyntheticDefaultImports in a consuming application does not appear to resolve the warning.
- Update Add TS definition for Configuration.setConfigPath. This definition was missing, keeping TS projects from being able to define a config path. However, the inline require() usage is ultimately the root of the issue.